Data Surveyor Added
This is a file which will write out all of the metadata within the raw zarr files to a file. I'll be adapting post-process.py soon to expect this metadata files. This will dramatically speed up the processing time for post-process. It will also provide a data check on all images.